created new repo
Had to create a new repo because I completed my homework on a fork after making an error. Forking my repo was something I tried to fix the issue but Netlify won't allow me to host a repo that is a branch without paying for a domain. All my commits can be provided if required to prove this is my own work.